+++ __tests__/TokenStorage.test.js
@@ -4,6 +4,7 @@ import TokenGeneratorMock from '../__mocks__/TokenGeneratorMock';
 import oauthClientCredentialsMock from '../__mocks__/oauthClientCredentials.json';
 import refreshedCredentials from '../__mocks__/refreshedCredentials.json';
 import Storage from '../__mocks__/mockStorage';
+import { NOW_TIMESTAMP_MOCK } from '../setupJest';
 
 global.FormData = require('form-data');
 
@@ -63,20 +64,36 @@ describe('Token storage tests', () => {
 
     expect(generatedToken).toBeInstanceOf(Promise);
 
+    const expectedStoredAccessTokenObject=  Object.assign(
+      {},
+      oauthClientCredentialsMock,
+      {
+        expires_at: 1487080308,
+      }
+    );
+
+    const expectedStoredRefreshTokenObject=  Object.assign(
+      {},
+      refreshedCredentials,
+      {
+        expires_at: 1487089508,
+      }
+    );
+
     return Promise.all([
       expect(typeof generatedToken).toBe('object'),
       expect(generatedToken.then(a => a.access_token)).resolves.toEqual(
-        oauthClientCredentialsMock.access_token
+        expectedStoredAccessTokenObject.access_token
       ),
     ])
       .then(() =>
         Promise.all([
           expect(oauth.hasAccessToken()).resolves.toBe(true),
           expect(oauth.getAccessToken()).resolves.toEqual(
-            oauthClientCredentialsMock.access_token
+            expectedStoredAccessTokenObject.access_token
           ),
           expect(oauth.getAccessTokenObject()).resolves.toEqual(
-            oauthClientCredentialsMock
+            expectedStoredAccessTokenObject
           ),
         ])
       )
@@ -84,17 +101,17 @@ describe('Token storage tests', () => {
         Promise.all([
           expect(
             oauth.refreshToken().then(a => a.access_token)
-          ).resolves.toEqual(refreshedCredentials.access_token),
+          ).resolves.toEqual(expectedStoredRefreshTokenObject.access_token),
         ])
       )
       .then(() =>
         Promise.all([
           expect(oauth.hasAccessToken()).resolves.toBe(true),
           expect(oauth.getAccessToken()).resolves.toEqual(
-            refreshedCredentials.access_token
+            expectedStoredRefreshTokenObject.access_token
           ),
           expect(oauth.getAccessTokenObject()).resolves.toEqual(
-            refreshedCredentials
+            expectedStoredRefreshTokenObject
           ),
         ])
       );
@@ -106,5 +123,18 @@ describe('Token storage tests', () => {
     await oauth._storeAccessToken('coucou');
     const actual = await oauth.getAccessTokenObject();
     expect(actual).toBeNull();
+  });
+
+  test('We can have the remaining validity time for a given token', async () => {
+    const oauth = new TokenStorage(tokenGeneratorMock, new Storage());
+    fetchMock.once(() => true, oauthClientCredentialsMock);
+
+    await oauth.generateToken({
+      grant_type: 'client_credentials',
+    });
+
+    expect(oauth.getCurrentTokenExpiresIn()).resolves.toEqual(3600);
+    await oauth.refreshToken();
+    expect(oauth.getCurrentTokenExpiresIn()).resolves.toEqual(12800);
   });
 });